Moved map-context-menu
